Move save function to io_utils
