+++ src/agent_factory/utils/__init__.py
@@ -1,9 +1,10 @@
 from .artifact_validation import clean_python_code_with_autoflake
-from .io_utils import setup_output_directory
+from .io_utils import save_agent_outputs, setup_output_directory
 from .logging import logger
 
 __all__ = [
     "clean_python_code_with_autoflake",
     "setup_output_directory",
     "logger",
+    "save_agent_outputs",
 ]
